+++ src/internal/kopia/snapshot_manager.go
@@ -24,6 +24,9 @@ const (
 	// (permalinks)
 	// [1] https://github.com/kopia/kopia/blob/05e729a7858a6e86cb48ba29fb53cb6045efce2b/cli/command_snapshot_create.go#L169
 	userTagPrefix = "tag:"
+
+	// Tag key applied to checkpoints (but not completed snapshots) in kopia.
+	checkpointTagKey = "checkpoint"
 )
 
 type Reason struct {
@@ -66,30 +69,6 @@ type snapshotManager interface {
 	LoadSnapshots(ctx context.Context, ids []manifest.ID) ([]*snapshot.Manifest, error)
 }
 
-type OwnersCats struct {
-	ResourceOwners map[string]struct{}
-	ServiceCats    map[string]ServiceCat
-}
-
-type ServiceCat struct {
-	Service  path.ServiceType
-	Category path.CategoryType
-}
-
-// MakeServiceCat produces the expected OwnersCats.ServiceCats key from a
-// path service and path category, as well as the ServiceCat value.
-func MakeServiceCat(s path.ServiceType, c path.CategoryType) (string, ServiceCat) {
-	return serviceCatString(s, c), ServiceCat{s, c}
-}
-
-// TODO(ashmrtn): Remove in a future PR.
-//
-//nolint:unused
-//lint:ignore U1000 will be removed in future PR.
-func serviceCatTag(p path.Path) string {
-	return serviceCatString(p.Service(), p.Category())
-}
-
 func serviceCatString(s path.ServiceType, c path.CategoryType) string {
 	return s.String() + c.String()
 }
@@ -102,33 +81,6 @@ func serviceCatString(s path.ServiceType, c path.CategoryType) string {
 // empty string.
 func makeTagKV(k string) (string, string) {
 	return userTagPrefix + k, defaultTagValue
-}
-
-// tagsFromStrings returns a map[string]string with tags for all ownersCats
-// passed in. Currently uses placeholder values for each tag because there can
-// be multiple instances of resource owners and categories in a single snapshot.
-// TODO(ashmrtn): Remove in future PR.
-//
-//nolint:unused
-//lint:ignore U1000 will be removed in future PR.
-func tagsFromStrings(oc *OwnersCats) map[string]string {
-	if oc == nil {
-		return map[string]string{}
-	}
-
-	res := make(map[string]string, len(oc.ServiceCats)+len(oc.ResourceOwners))
-
-	for k := range oc.ServiceCats {
-		tk, tv := makeTagKV(k)
-		res[tk] = tv
-	}
-
-	for k := range oc.ResourceOwners {
-		tk, tv := makeTagKV(k)
-		res[tk] = tv
-	}
-
-	return res
 }
 
 // getLastIdx searches for manifests contained in both foundMans and metas